+++ src/math/ft_round.c
@@ -6,7 +6,7 @@
 /*   By: jaicastr <jaicastr@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/16 17:13:42 by jaicastr          #+#    #+#             */
-/*   Updated: 2026/01/19 04:38:34 by jaicastr         ###   ########.fr       */
+/*   Updated: 2026/03/15 15:39:48 by jaicastr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,8 @@ inline float	ft_floorf(float x)
 	long long	i;
 
 	i = (long long)x;
-	return ((float)(((i > x) * (i - 1)) | ((i <= x) * i)));
+	return ((float)((((float)i > x) * (i - 1))
+		| (((float)i <= x) * i)));
 }
 
 __attribute__((__always_inline__, const))
@@ -27,7 +28,8 @@ inline float	ft_ceilf(float x)
 	long long	i;
 
 	i = (long long)x;
-	return ((float)(((i > x) * (i + 1)) | ((i <= x) * i)));
+	return ((float)((((float)i > x) * (i + 1))
+		| (((float)i <= x) * i)));
 }
 
 __attribute__((__always_inline__, const))
@@ -44,8 +46,8 @@ inline float	ft_roundf(float x, t_u8 n)
 		p10 *= 10.0f;
 	sign = (x < 0);
 	offset = 0.5f - (float)sign;
-	return ((ft_floorf(x * p10 + offset) * (1 - sign)
-			+ ft_ceilf(x * p10 + offset) * sign) / p10);
+	return ((ft_floorf(x * p10 + offset) * (float)(1 - sign)
+			+ ft_ceilf(x * p10 + offset) * (float)sign) / p10);
 }
 
 __attribute__((__always_inline__, const))
